Implement item respawn in multiplayer mode.
Created a new request: ItemRespawnRequest. This would allow the server to be the authority on whether item has been taken or not. Current implementation is such that only the master client manage handle item respawn. The information is then updated at the server side with a ItemRespawnRequest. On the other hand, slave client will just obey the authority and respawn item whenever it is truly available.