Fix monster position inconsistency in multiplayer.
Added a check to make prevent monster from continuously chasing a target if the
target is too far away (eg. player died and respawned) while the client is
awaiting server's confirmation on stopping chase. If the target is indeed
too far, it will be set to the monster's current position.