Fix monster moving out of center of isle.
Reworked the chasing algorithm to make sure that monster will always stick to the center of isle. The map walls are changed a little to ensure all width of path are the same in order for the algorithm to work properly.

The size of monster is also increased to cover up scenario where monster isn't moving because the player is too near it.